+++ sql-plugin/src/main/spark341db/scala/org/apache/spark/rapids/shims/api/python/ShimBasePythonRunner.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021-2023, NVIDIA CORPORATION.
+ * Copyright (c) 2023, NVIDIA CORPORATION.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,46 +15,29 @@
  */
 
 /*** spark-rapids-shim-json-lines
-{"spark": "320"}
-{"spark": "321"}
-{"spark": "321cdh"}
-{"spark": "321db"}
-{"spark": "322"}
-{"spark": "323"}
-{"spark": "324"}
-{"spark": "330"}
-{"spark": "330cdh"}
-{"spark": "330db"}
-{"spark": "331"}
-{"spark": "332"}
-{"spark": "332cdh"}
-{"spark": "332db"}
-{"spark": "333"}
-{"spark": "340"}
-{"spark": "341"}
+{"spark": "341db"}
 spark-rapids-shim-json-lines ***/
 package org.apache.spark.rapids.shims.api.python
 
 import java.io.DataInputStream
-import java.net.Socket
 import java.util.concurrent.atomic.AtomicBoolean
 
 import org.apache.spark.{SparkEnv, TaskContext}
-import org.apache.spark.api.python.BasePythonRunner
+import org.apache.spark.api.python.{BasePythonRunner, PythonWorker}
 
 abstract class ShimBasePythonRunner[IN, OUT](
     funcs : scala.Seq[org.apache.spark.api.python.ChainedPythonFunctions],
     evalType : scala.Int, argOffsets : scala.Array[scala.Array[scala.Int]]
-) extends BasePythonRunner[IN, OUT](funcs, evalType, argOffsets) {
+) extends BasePythonRunner[IN, OUT](funcs, evalType, argOffsets, None) {
   protected abstract class ShimReaderIterator(
     stream: DataInputStream,
-    writerThread: WriterThread,
+    writer: Writer,
     startTime: Long,
     env: SparkEnv,
-    worker: Socket,
+    worker: PythonWorker,
     pid: Option[Int],
     releasedOrClosed: AtomicBoolean,
     context: TaskContext
-  ) extends ReaderIterator(stream, writerThread, startTime, env, worker, pid,
+  ) extends ReaderIterator(stream, writer, startTime, env, worker, pid,
     releasedOrClosed, context)
 }
