+++ sql-plugin/src/main/spark341db/scala/org/apache/spark/sql/rapids/execution/python/shims/GpuGroupUDFArrowPythonRunner.scala
@@ -18,14 +18,11 @@
  */
 
 /*** spark-rapids-shim-json-lines
-{"spark": "321db"}
-{"spark": "330db"}
-{"spark": "332db"}
+{"spark": "341db"}
 spark-rapids-shim-json-lines ***/
 package org.apache.spark.sql.rapids.execution.python.shims
 
 import java.io.DataOutputStream
-import java.net.Socket
 
 import ai.rapids.cudf._
 import com.nvidia.spark.rapids._
@@ -64,13 +61,13 @@ class GpuGroupUDFArrowPythonRunner(
   extends GpuPythonRunnerBase[ColumnarBatch](funcs, evalType, argOffsets)
     with GpuPythonArrowOutput {
 
-  protected override def newWriterThread(
+  protected override def newWriter(
       env: SparkEnv,
-      worker: Socket,
+      worker: PythonWorker,
       inputIterator: Iterator[ColumnarBatch],
       partitionIndex: Int,
-      context: TaskContext): WriterThread = {
-    new WriterThread(env, worker, inputIterator, partitionIndex, context) {
+      context: TaskContext): Writer = {
+    new Writer(env, worker, inputIterator, partitionIndex, context) {
 
       protected override def writeCommand(dataOut: DataOutputStream): Unit = {
 
@@ -84,7 +81,8 @@ class GpuGroupUDFArrowPythonRunner(
         PythonUDFRunner.writeUDFs(dataOut, funcs, argOffsets)
       }
 
-      protected override def writeIteratorToStream(dataOut: DataOutputStream): Unit = {
+      override def writeNextInputToStream(dataOut: DataOutputStream): Boolean = {
+        var wrote = false
         // write out number of columns
         Utils.tryWithSafeFinally {
           val builder = ArrowIPCWriterOptions.builder()
@@ -102,6 +100,7 @@ class GpuGroupUDFArrowPythonRunner(
               }
           }
           while(inputIterator.hasNext) {
+            wrote = false
             val writer = {
               // write 1 out to indicate there is more to read
               dataOut.writeInt(1)
@@ -113,6 +112,7 @@ class GpuGroupUDFArrowPythonRunner(
             withResource(new NvtxRange("write python batch", NvtxColor.DARK_GREEN)) { _ =>
               // The callback will handle closing table and releasing the semaphore
               writer.write(table)
+              wrote = true
             }
             writer.close()
             dataOut.flush()
@@ -125,6 +125,7 @@ class GpuGroupUDFArrowPythonRunner(
           dataOut.writeInt(0)
           dataOut.flush()
         }
+        wrote
       }
     }
   }
